feat: add search page - rename skills page
