+++ src/utils/router.tsx
@@ -3,6 +3,7 @@ import { Auth } from '../components/Auth';
 import { Effort } from '../components/Effort';
 import { Profile } from '../components/Profile';
 import { Search } from '../components/Search';
+import { Skills } from '../components/Skills';
 
 export type Route = keyof typeof routes;
 
@@ -10,6 +11,7 @@ export const routes = {
 	auth: <Auth />,
 	profile: <Profile />,
 	effort: <Effort />,
+	skills: <Skills />,
 	search: <Search />,
 };
